Move helper functions to common-utils.
