+++ commands/miscinfo/miscinfo_handler.py
@@ -1,5 +1,4 @@
 from commands.miscinfo.action.get_ids_from_szslibrary import get_wbz_ids
-
 from commands.miscinfo.action.edit_ids_on_page import add_ids_from_file
 
 def handle_command(args):
@@ -9,7 +8,6 @@ def handle_command(args):
         end_id = args["end_id"]
         file_path = args["dump_file_path"]
         get_wbz_ids(start_id, end_id, file_path)
-
     elif action == "write_ids":
         file_path = args["dump_file_path"]
         add_ids_from_file(file_path)
